Rework responsive layout
Mobile is now the default - switch to desktop layout on larger screens,
rather than the reverse.

On desktop, use a fixed size centered block, rather than scaling to
content. Makes it visually more consistent with mobile.

Links at the bottom of the page now stick to the bottom of the page (or
content block on desktop), rather than being centered vertically with the
main content. Similarly, header on the sites page sticks to the top.
Gives more consistent positioning across pages - and just looks better.