Rework redirect handling
The spec explicitly allows redirecting to another location - including a
different domain - to serve the support resource.

Rather than failing the site if it redirects to a different scheme, domain,
or path, add a warning (not shown to the user) and continue.

Only allow the final resource to use HTTP or HTTPS schemes, however - don't
want to allow other random schemes.